fix(fuzz): retarget the fuzz harness after removing the specifier parser
The version-matrix rewrite deleted parse_version/satisfies, which the fuzz
harness imported — breaking the ClusterFuzzLite build (import error at compile
time). The new version-matrix path is stdlib tomllib + an anchored regex +
packaging, with no hand-rolled parsing left worth fuzzing.

Retarget fuzzing at the suppression comment matcher, which is the remaining
untrusted-text parser (and security-relevant: it detects # nosec CVE
suppressions):

- add tests/fuzz/fuzz_suppression.py exercising _match_suppression + nosec_cves;
- delete tests/fuzz/fuzz_version_matrix.py and the unreferenced, also-broken
  fuzzers/version_matrix_fuzzer.py duplicate;
- update the .clusterfuzzlite/build.sh comment to match the new target.
